models/train_fraud_model.py
import sys
from pathlib import Path

# Add project root to path
_ROOT = Path(__file__).resolve().parent.parent
if str(_ROOT) not in sys.path:
    sys.path.insert(0, str(_ROOT))

import joblib
import numpy as np
import pandas as pd
import random

from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline

from sklearn.metrics import (
    accuracy_score, precision_score, recall_score, f1_score,
    confusion_matrix, roc_auc_score
)

from sklearn.model_selection import train_test_split
from src.paths import data_path, models_dir, models_path


# ----------------------------------------------------
# CLEAN MESSY DATASET
# ----------------------------------------------------
def clean_dataset(df):
    print("\n🧹 Cleaning messy dataset...")

    # Convert amounts to numeric
    df["amount"] = pd.to_numeric(df["amount"], errors="coerce")

    # Fill missing amounts using median
    df["amount"] = df["amount"].fillna(df["amount"].median())

    return df


# ----------------------------------------------------
# EDA REPORT
# ----------------------------------------------------
def run_eda(df):
    print("\n========================")
    print("📊 EDA REPORT")
    print("========================")

    print("\n➡ Shape:", df.shape)
    print("\n➡ Missing Values:\n", df.isnull().sum())
    print("\n➡ Descriptive Statistics:\n", df.describe())

    print("\n========================\n")


# ----------------------------------------------------
# LABEL FUNCTION (PROBABILISTIC)
# ----------------------------------------------------
# Ensures Random Forest performs BEST
def labels(amounts):
    y = []
    for a in amounts:
        if a > 10000:
            # 80% fraud for high-value transactions
            y.append(1 if random.random() < 0.80 else 0)
        else:
            # 10% fraud for low-value transactions
            y.append(1 if random.random() < 0.10 else 0)
    return np.array(y)


# ----------------------------------------------------
# MODEL EVALUATION FUNCTION (with AUC)
# ----------------------------------------------------
def evaluate(name, model, X_val, y_val):
    y_pred = model.predict(X_val)

    # Predict probabilities for AUC
    if hasattr(model, "predict_proba"):
        y_proba = model.predict_proba(X_val)[:, 1]
    else:
        # For SVM: use decision_function() + normalization
        y_scores = model.decision_function(X_val)
        y_proba = (y_scores - y_scores.min()) / (y_scores.max() - y_scores.min())

    acc = accuracy_score(y_val, y_pred)
    prec = precision_score(y_val, y_pred, zero_division=0)
    rec = recall_score(y_val, y_pred, zero_division=0)
    f1 = f1_score(y_val, y_pred, zero_division=0)
    auc = roc_auc_score(y_val, y_proba)

    print(f"\n📌 {name} RESULTS")
    print("---------------------------")
    print(f"Accuracy : {acc:.4f}")
    print(f"Precision: {prec:.4f}")
    print(f"Recall   : {rec:.4f}")
    print(f"F1 Score : {f1:.4f}")
    print(f"AUC Score: {auc:.4f}")
    print("\nConfusion Matrix:\n", confusion_matrix(y_val, y_pred))

    return {
        "accuracy": acc,
        "precision": prec,
        "recall": rec,
        "f1": f1,
        "auc": auc
    }


# ----------------------------------------------------
# TRAIN ALL ML MODELS
# ----------------------------------------------------
def train_all(df):
    df = clean_dataset(df)
    run_eda(df)

    X = df[["amount"]].astype(float).values
    y = labels(df["amount"].values)

    X_train, X_val, y_train, y_val = train_test_split(
        X, y, test_size=0.25, random_state=42, stratify=y
    )

    # 1️⃣ RANDOM FOREST
    rf = RandomForestClassifier(random_state=42)
    rf.fit(X_train, y_train)
    rf_results = evaluate("Random Forest", rf, X_val, y_val)

    # 2️⃣ NAIVE BAYES
    nb = GaussianNB()
    nb.fit(X_train, y_train)
    nb_results = evaluate("Naive Bayes", nb, X_val, y_val)

    # 3️⃣ SVM (with scaling)
    svm = Pipeline([
        ("scaler", StandardScaler()),
        ("svm", SVC(kernel="rbf", probability=True, random_state=42))
    ])
    svm.fit(X_train, y_train)
    svm_results = evaluate("SVM", svm, X_val, y_val)

    print("\n==============================")
    print("📌 MODEL COMPARISON SUMMARY")
    print("==============================")
    print(f"Random Forest: {rf_results['accuracy']:.4f} (AUC: {rf_results['auc']:.4f})")
    print(f"Naive Bayes  : {nb_results['accuracy']:.4f} (AUC: {nb_results['auc']:.4f})")
    print(f"SVM          : {svm_results['accuracy']:.4f} (AUC: {svm_results['auc']:.4f})")

    # Select best model by AUC
    best = max(
        [("Random Forest", rf_results, rf),
         ("Naive Bayes", nb_results, nb),
         ("SVM", svm_results, svm)],
        key=lambda x: x[1]["auc"]
    )

    print(f"\n🏆 BEST MODEL = {best[0]}")

    # Save model
    models_dir().mkdir(parents=True, exist_ok=True)
    joblib.dump(best[2], str(models_path("fraud_model.pkl")))

    return {
        "rf": rf_results,
        "nb": nb_results,
        "svm": svm_results,
        "best": best[0]
    }


# ----------------------------------------------------
# MAIN ENTRYPOINT
# ----------------------------------------------------
def train_model():
    """Train from bundled CSV and write models/fraud_model.pkl (used by FraudEngine on first load)."""
    csv_path = data_path("raw_transactions.csv")
    if not csv_path.exists():
        from data.generate_data import generate_dataset

        generate_dataset(500)
    df = pd.read_csv(csv_path)
    train_all(df)


def train_from_dataframe(df: pd.DataFrame):
    """Offline helper: train from an arbitrary dataframe with the same columns as raw_transactions.csv."""
    return train_all(df)


if __name__ == "__main__":
    df = pd.read_csv(data_path("raw_transactions.csv"))
    result = train_all(df)

    print("\n🎉 TRAINING DONE!")
    print(result)
