requirements.txt
streamlit
pandas
numpy
scikit-learn
spacy==3.7.4
joblib
faker
python-dotenv
openai
requests
yagmail

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
